requirement.txt

Flask==2.0.2
gast==0.5.3
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.42.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
Markdown==3.3.6
MarkupSafe==2.0.1
mtcnn==0.1.1
numpy==1.19.5
opencv-python==4.1.0.25
Pillow==8.4.0
protobuf==3.19.1
pyodbc==4.0.32
PyYAML==6.0
requests==2.26.0
scipy==1.5.4
six==1.16.0
SQLAlchemy==1.4.27
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
typing-extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
